Imports restructurized
