+++ alcor/services/data_access/__init__.py
@@ -1,5 +1,5 @@
 from .creating import insert
+from .reading import fetch_query
 from .statements import (select_statement,
                          insert_statement,
                          model_insert_statement)
-from .reading import fetch_query
